[style] Controller, Service, Model, View 디렉토리 구분
