+++ src/main/java/lotto/model/Lotto.java
@@ -1,34 +1,24 @@
 package lotto.model;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.IntStream;
 
 public class Lotto {
-    private final Numbers numbers;
+    private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
-        this.numbers = Numbers.valueOf(numbers);
-    }
-
-    public Lotto(Numbers numbers) {
         this.numbers = numbers;
     }
 
-    public Numbers getNumbers() {
-        return numbers;
+    public List<Integer> getNumbers() {
+        return Collections.unmodifiableList(numbers);
     }
 
-    public MatchNumber matchNumbers(Numbers winningNumbers, Number bonusNumber) {
-        return new MatchNumber(getMatchCount(winningNumbers), getMatchBonus(bonusNumber));
-    }
-
-    private boolean getMatchBonus(Number bonusNumber) {
-        return numbers.contains(bonusNumber);
-    }
-
-    private int getMatchCount(Numbers winningNumbers) {
-        return (int) winningNumbers.getNumberList().stream()
+    public int matchNumbers(List<Integer> winningNumbers) {
+        return (int) IntStream.range(0, winningNumbers.size())
                 .parallel()
-                .filter(number -> getMatchBonus(number))
+                .filter(i -> numbers.contains(winningNumbers.get(i)))
                 .count();
     }
 }
