+++ src/main/java/lotto/model/LottoOutlet.java
@@ -6,27 +6,13 @@ import java.util.stream.Stream;
 
 public class LottoOutlet {
     private static final int LOTTO_PRICE = 1000;
-    private static final String LACK_MONEY_MESSAGE = "해당 금액으로는 1장도 구매하실 수 없습니다.";
-    private static final NumbersGenerator numbersGenerator = new RandomNumbersGeneratorImpl();
 
     public static int lottoCount(int money) {
-        int lottoCount = calculateLottoCount(money);
-        validateLottoCount(lottoCount);
-        return lottoCount;
-    }
-
-    private static int calculateLottoCount(int money) {
         return money / LOTTO_PRICE;
     }
 
-    private static void validateLottoCount(int lottoCount) {
-        if (lottoCount < 1) {
-            throw new IllegalArgumentException(LACK_MONEY_MESSAGE);
-        }
-    }
-
     public static Lotto generateLotto() {
-        return new Lotto(numbersGenerator.generate());
+        return new Lotto(List.of(1, 2, 3, 4, 5, 6));
     }
 
     public static List<Lotto> generateLottos(int count) {
@@ -35,17 +21,21 @@ public class LottoOutlet {
                 .collect(Collectors.toList());
     }
 
-    public static int getWinnings(List<MatchNumber> matchNumbers) {
-        return matchNumbers.stream()
-                .mapToInt(matchNumber -> Rank.getWinningsAmount(matchNumber))
+    public static int getWinning(int matchCount) {
+        return Winnings.getWinningsAmount(matchCount);
+    }
+
+    public static int getWinnings(List<Integer> matchCounts) {
+        return matchCounts.stream()
+                .mapToInt(Winnings::getWinningsAmount)
                 .sum();
     }
 
-    public static int getWinning(MatchNumber matchNumber) {
-        return Rank.getWinningsAmount(matchNumber);
-    }
-
-    public static int getRank(MatchNumber matchNumber) {
-        return Rank.getRank(matchNumber);
+    public static int getWinnings(List<Lotto> Lottos, List<Integer> winningNumbers) {
+        return Lottos.stream()
+                .mapToInt(lotto -> lotto.matchNumbers(winningNumbers))
+                .boxed()
+                .mapToInt(Winnings::getWinningsAmount)
+                .sum();
     }
 }
