+++ src/main/java/lotto/service/LottoService.java
@@ -2,29 +2,16 @@ package lotto.service;
 
 import lotto.model.Lotto;
 import lotto.model.LottoOutlet;
-import lotto.model.Numbers;
 import lotto.model.ResultLotto;
 
 import java.util.List;
 
 public class LottoService {
-    private final ResultLotto resultLotto;
-
-    public LottoService() {
-        this.resultLotto = new ResultLotto();
-    }
-
-    public ResultLotto buyLottos(int seedMoney) {
+    public ResultLotto youOnlyLiveOnce(int seedMoney) {
         int lottoCount = LottoOutlet.lottoCount(seedMoney);
         List<Lotto> buyLottos = LottoOutlet.generateLottos(lottoCount);
+        int totalWinningAmount = LottoOutlet.getWinnings(buyLottos, List.of(11,12,13,4,5,6));
 
-        resultLotto.recordBuyLottos(buyLottos);
-        return resultLotto;
-    }
-
-    public ResultLotto checkLottoResult(List<Integer> winningNumbers, int bonusNumber) {
-        resultLotto.recordWinningNumbers(Numbers.valueOf(winningNumbers), bonusNumber);
-
-        return resultLotto;
+        return new ResultLotto(seedMoney, buyLottos, totalWinningAmount);
     }
 }
